Extract focus-icon GFX browser, menubar, toolbar; final docs true-up
