+++ hoi4_content_maker.py
@@ -98,6 +98,9 @@ from hoi4cm.ui.canvas import CanvasMixin
 from hoi4cm.ui.mod_loading import ModLoadingMixin
 from hoi4cm.ui.effects_panel import EffectsMixin
 from hoi4cm.ui.settings_dialog import open_settings
+from hoi4cm.ui.gfx_browser import open_focus_icon_browser
+from hoi4cm.ui.menubar import build_menubar
+from hoi4cm.ui.toolbar import build_toolbar_row2
 from hoi4cm.ui import (
     BG_CARD,
     BG_DARK,
@@ -120,7 +123,6 @@ from hoi4cm.ui import (
     ORANGE,
     PREREQ_COL,
     RED,
-    SEL_BG,
     TEXT,
     TEXT_DIM,
     XGRID,
@@ -128,7 +130,6 @@ from hoi4cm.ui import (
     YGRID,
     Tooltip,
     _safe_after,
-    _safe_after_idle,
     make_progress,
     progress_modal,
     run_bg,
@@ -190,7 +191,6 @@ import copy
 import json
 import re
 import subprocess
-import threading
 import time
 
 
@@ -399,805 +399,10 @@ class App(CanvasMixin, ModLoadingMixin, EffectsMixin, tk.Tk):
         self._extra_trees = []  # list of dicts: {type, file_path, tree_id, cfp_x, cfp_y, shared_focuses, joint_focuses, country_tag, focus_ids}
         toolbar = tk.Frame(self, bg=BG_DARK)
         toolbar.pack(fill="x")
-        self._build_menubar(toolbar)
-        self._build_toolbar_row2(toolbar)
+        build_menubar(self, toolbar)
+        build_toolbar_row2(self, toolbar)
         self._build_keybinds()
         self._build_layout()
-
-    def _build_menubar(self, toolbar):
-        """Build menu bar with File/Edit/View/Tools dropdowns."""
-        # ── MENU BAR ──────────────────────────────────────────────
-        menubar = tk.Frame(toolbar, bg="#080b10", height=30)
-        menubar.pack(fill="x")
-        menubar.pack_propagate(False)
-        tk.Frame(menubar, bg=BORDER_G, height=1).place(relx=0, rely=1, relwidth=1)
-
-        # Branding
-        tk.Label(
-            menubar,
-            text=tr("app.brand", "HOI4 CONTENT MAKER"),
-            bg="#080b10",
-            fg=TEXT_DIM,
-            font=("Helvetica", 9, "bold"),
-            padx=12,
-        ).pack(side="left")
-        tk.Frame(menubar, bg=BORDER_G, width=1, height=18).pack(side="left", padx=2)
-
-        # ── Menu helpers ──────────────────────────────────────────
-        self._open_menu_win = [None]  # [current Toplevel or None]
-        self._open_menu_btn = [None]  # [button that opened it]
-
-        def _close_menu():
-            w = self._open_menu_win[0]
-            b = self._open_menu_btn[0]
-            if w:
-                try:
-                    w.destroy()
-                except Exception:
-                    pass
-            self._open_menu_win[0] = None
-            self._open_menu_btn[0] = None
-            if b:
-                try:
-                    b.config(bg="#080b10", fg=TEXT_DIM)
-                except Exception:
-                    pass
-
-        def _menu_btn(parent, label, items):
-            btn = tk.Button(
-                parent,
-                text=label,
-                bg="#080b10",
-                fg=TEXT_DIM,
-                font=("Helvetica", 9, "bold"),
-                relief="flat",
-                padx=10,
-                pady=0,
-                cursor="hand2",
-                bd=0,
-                activebackground="#1a2030",
-                activeforeground=TEXT,
-            )
-            btn.pack(side="left")
-
-            def _open(b=btn):
-                # If this menu is already open, close it
-                if self._open_menu_btn[0] is b:
-                    _close_menu()
-                    return
-                _close_menu()
-                b.config(bg="#1a2030", fg=TEXT)
-                # Build a Toplevel dropdown
-                b.update_idletasks()
-                rx = b.winfo_rootx()
-                ry = b.winfo_rooty() + b.winfo_height()
-                drop = tk.Toplevel(self)
-                drop.wm_overrideredirect(True)
-                drop.configure(bg="#0d1218")
-                drop.attributes("-topmost", True)
-                # build contents
-                inner = tk.Frame(
-                    drop,
-                    bg="#0d1218",
-                    highlightthickness=1,
-                    highlightbackground=BORDER_G,
-                )
-                inner.pack(fill="both", expand=True)
-                min_w = 340
-                for item in items:
-                    if item is None:
-                        tk.Frame(inner, bg=BORDER_G, height=1).pack(
-                            fill="x", padx=0, pady=2
-                        )
-                    elif isinstance(item, str):
-                        tk.Label(
-                            inner,
-                            text=item.upper(),
-                            bg="#0d1218",
-                            fg=TEXT_DIM,
-                            font=("Helvetica", 8, "bold"),
-                            anchor="w",
-                            padx=16,
-                            pady=3,
-                        ).pack(fill="x")
-                    else:
-                        lbl_i, cmd_i, *rest = item
-                        kbd = rest[0] if len(rest) > 0 else ""
-                        tip_txt = rest[1] if len(rest) > 1 else ""
-                        item_bg_n = "#0d1218"
-                        item_bg_h = "#141c2a"
-                        row_f = tk.Frame(inner, bg=item_bg_n)
-                        row_f.pack(fill="x")
-
-                        def _make_cmd(c=cmd_i):
-                            def _do():
-                                _close_menu()
-                                if c:
-                                    self.after(10, c)
-
-                            return _do
-
-                        # main label + kbd shortcut row
-                        top_row = tk.Frame(row_f, bg=item_bg_n)
-                        top_row.pack(fill="x")
-                        ib = tk.Button(
-                            top_row,
-                            text=lbl_i,
-                            command=_make_cmd(),
-                            bg=item_bg_n,
-                            fg=TEXT_DIM,
-                            font=("Helvetica", 9, "bold"),
-                            relief="flat",
-                            anchor="w",
-                            padx=16,
-                            pady=4,
-                            cursor="hand2",
-                            bd=0,
-                            activebackground=item_bg_h,
-                            activeforeground=TEXT,
-                        )
-                        ib.pack(side="left", fill="x", expand=True)
-                        if kbd:
-                            tk.Label(
-                                top_row,
-                                text=kbd,
-                                bg=item_bg_n,
-                                fg="#3a4a60",
-                                font=("Courier", 8),
-                                padx=12,
-                            ).pack(side="right")
-                        # tooltip description line (dim, indented)
-                        if tip_txt:
-                            tip_lbl = tk.Label(
-                                row_f,
-                                text=tip_txt,
-                                bg=item_bg_n,
-                                fg="#3d5068",
-                                font=("Helvetica", 8),
-                                anchor="w",
-                                padx=26,
-                                pady=0,
-                            )
-                            tip_lbl.pack(fill="x")
-                        else:
-                            tip_lbl = None
-                        # hover: highlight entire row including tip
-                        all_widgets = [row_f, top_row, ib] + (
-                            [tip_lbl] if tip_lbl else []
-                        )
-
-                        def _enter(e, ws=all_widgets, kl=top_row, tl=tip_lbl, ib_=ib):
-                            for w in ws:
-                                w.config(bg=item_bg_h)
-                            ib_.config(fg=TEXT)
-                            if tl:
-                                tl.config(fg="#6b849a")
-
-                        def _leave(e, ws=all_widgets, tl=tip_lbl, ib_=ib):
-                            for w in ws:
-                                w.config(bg=item_bg_n)
-                            ib_.config(fg=TEXT_DIM)
-                            if tl:
-                                tl.config(fg="#3d5068")
-
-                        for w in all_widgets:
-                            w.bind("<Enter>", _enter)
-                            w.bind("<Leave>", _leave)
-                # position and show
-                drop.update_idletasks()
-                dw = max(drop.winfo_reqwidth(), min_w)
-                dh = drop.winfo_reqheight()
-                drop.geometry(f"{dw}x{dh}+{rx}+{ry}")
-                # close on any click outside
-                drop.bind(
-                    "<FocusOut>",
-                    lambda e: self.after(
-                        100,
-                        lambda: _close_menu()
-                        if self._open_menu_win[0] is drop
-                        else None,
-                    ),
-                )
-                self._open_menu_win[0] = drop
-                self._open_menu_btn[0] = b
-
-            btn.config(command=_open)
-            btn.bind(
-                "<Enter>",
-                lambda e, b=btn: b.config(fg=TEXT)
-                if self._open_menu_btn[0] is not b
-                else None,
-            )
-            btn.bind(
-                "<Leave>",
-                lambda e, b=btn: b.config(fg=TEXT_DIM, bg="#080b10")
-                if self._open_menu_btn[0] is not b
-                else None,
-            )
-            return btn
-
-        # ── FILE MENU ─────────────────────────────────────────────
-        _menu_btn(
-            menubar,
-            tr("menu.file", "File"),
-            [
-                tr("menu.section.project", "Project"),
-                (
-                    tr("menu.new_tree", "New Tree"),
-                    self._new_tree_dialog,
-                    "Ctrl+N",
-                    tr(
-                        "menu.new_tree.tip",
-                        "Start fresh - set country tag and auto-prefix all focus IDs.",
-                    ),
-                ),
-                (
-                    tr("menu.save_project", "Save Project"),
-                    self._save,
-                    "Ctrl+S",
-                    tr(
-                        "menu.save_project.tip",
-                        "Save as .json so you can reopen and keep editing later.",
-                    ),
-                ),
-                (
-                    tr("menu.load_project", "Load Project"),
-                    self._load,
-                    "Ctrl+O",
-                    tr(
-                        "menu.load_project.tip",
-                        "Open a previously saved .json project file.",
-                    ),
-                ),
-                None,
-                tr("menu.section.recent", "Recent"),
-                *(
-                    [(tr("menu.no_recent_files", "  (no recent files)"), None)]
-                    if not getattr(MOD, "_recent_mods", [])
-                    else [
-                        (f"  {r}", lambda p=r: self._load_mod_path(p))
-                        for r in (getattr(MOD, "_recent_mods", []) or [])
-                    ]
-                ),
-                None,
-                tr("menu.section.import_export", "Import / Export"),
-                (
-                    tr("menu.export_txt", "Export .txt"),
-                    self._export,
-                    "Ctrl+E",
-                    tr(
-                        "menu.export_txt.tip",
-                        "Write HOI4-ready script to a .txt file for your mod folder.",
-                    ),
-                ),
-                (
-                    tr("menu.export_mod", "Export to Mod"),
-                    self._export,
-                    "Ctrl+Shift+E",
-                    tr(
-                        "menu.export_mod.tip",
-                        "Write directly into your loaded mod's national_focus folder.",
-                    ),
-                ),
-                (
-                    tr("menu.import_txt", "Import .txt"),
-                    self._import_txt,
-                    "",
-                    tr(
-                        "menu.import_txt.tip",
-                        "Read an existing HOI4 focus tree .txt and populate the canvas.",
-                    ),
-                ),
-                (
-                    tr("menu.import_drawio", "Import Draw.io"),
-                    self._import_drawio,
-                    "",
-                    tr(
-                        "menu.import_drawio.tip",
-                        "Convert a Draw.io diagram into focus nodes and prerequisite arrows.",
-                    ),
-                ),
-            ],
-        )
-
-        # ── EDIT MENU ─────────────────────────────────────────────
-        _menu_btn(
-            menubar,
-            tr("menu.edit", "Edit"),
-            [
-                (
-                    tr("menu.undo", "Undo"),
-                    self._undo,
-                    "Ctrl+Z",
-                    tr(
-                        "menu.undo.tip",
-                        "Revert the last canvas action (move, add, delete, edit).",
-                    ),
-                ),
-                None,
-                (
-                    tr("menu.duplicate_focus", "Duplicate Focus"),
-                    self._duplicate_focus,
-                    "",
-                    tr(
-                        "menu.duplicate_focus.tip",
-                        "Copy the selected focus - gets a _copy suffix, shifted one column right.",
-                    ),
-                ),
-                (
-                    tr("menu.bulk_rename_prefix", "Bulk Rename Prefix"),
-                    self._bulk_rename_dialog,
-                    "",
-                    tr(
-                        "menu.bulk_rename_prefix.tip",
-                        "Replace a prefix across all focus IDs, prerequisites and mutex links at once.",
-                    ),
-                ),
-                None,
-                (
-                    tr("menu.select_all", "Select All"),
-                    self._select_all_focuses,
-                    "Ctrl+A",
-                    tr(
-                        "menu.select_all.tip",
-                        "Enable multi-select and mark every focus on the canvas.",
-                    ),
-                ),
-                (
-                    tr("menu.delete_selected", "Delete Selected"),
-                    self._delete_selected,
-                    "Del",
-                    tr(
-                        "menu.delete_selected.tip",
-                        "Remove all multi-selected focuses after confirmation.",
-                    ),
-                ),
-            ],
-        )
-
-        # ── VIEW MENU ─────────────────────────────────────────────
-        _menu_btn(
-            menubar,
-            tr("menu.view", "View"),
-            [
-                (
-                    tr("menu.toggle_grid", "Toggle Grid"),
-                    self._toggle_grid,
-                    "G",
-                    tr(
-                        "menu.toggle_grid.tip",
-                        "Show or hide the background snap grid on the canvas.",
-                    ),
-                ),
-                (
-                    tr("menu.toggle_minimap", "Toggle Minimap"),
-                    self._toggle_minimap,
-                    "M",
-                    tr(
-                        "menu.toggle_minimap.tip",
-                        "Show or hide the minimap overview in the bottom-right corner.",
-                    ),
-                ),
-                (
-                    tr("menu.toggle_focus_list", "Toggle Focus List"),
-                    self._toggle_focus_list,
-                    "F",
-                    tr(
-                        "menu.toggle_focus_list.tip",
-                        "Collapse or expand the left-side focus list panel.",
-                    ),
-                ),
-                None,
-                (
-                    tr("menu.fit_all_focuses", "Fit All Focuses"),
-                    self._fit_all,
-                    "0",
-                    tr(
-                        "menu.fit_all_focuses.tip",
-                        "Zoom and pan so every focus on the canvas is visible at once.",
-                    ),
-                ),
-            ],
-        )
-
-        # ── TOOLS MENU ────────────────────────────────────────────
-        _menu_btn(
-            menubar,
-            tr("menu.tools", "Tools"),
-            [
-                (
-                    tr("menu.national_spirit_builder", "National Spirit Builder"),
-                    self._national_spirit_wizard,
-                    "",
-                    tr(
-                        "menu.national_spirit_builder.tip",
-                        "Create ideas/spirits with a modifier editor and live HOI4 preview.",
-                    ),
-                ),
-                (
-                    tr("menu.dynamic_modifier", "Dynamic Modifier"),
-                    self._dyn_mod_wizard,
-                    "",
-                    tr(
-                        "menu.dynamic_modifier.tip",
-                        "Build add_dynamic_modifier effects with variable-driven scaling.",
-                    ),
-                ),
-                (
-                    tr("menu.decision_maker", "Decision Maker"),
-                    self._decision_wizard,
-                    "",
-                    tr(
-                        "menu.decision_maker.tip",
-                        "Build decisions and decision categories with GFX placement editor.",
-                    ),
-                ),
-                (
-                    tr("menu.event_maker", "Event Maker"),
-                    self._event_wizard,
-                    "",
-                    tr(
-                        "menu.event_maker.tip",
-                        "Build country_event / news_event blocks with options and live preview.",
-                    ),
-                ),
-                None,
-                (
-                    tr("menu.validate_tree", "Validate Tree"),
-                    self._validate_tree,
-                    "",
-                    tr(
-                        "menu.validate_tree.tip",
-                        "Check for broken prerequisites, missing effects, and bad GFX references.",
-                    ),
-                ),
-                (
-                    tr("menu.load_mod", "Load Mod"),
-                    self._load_mod,
-                    "",
-                    tr(
-                        "menu.load_mod.tip",
-                        "Point to your mod root folder to browse GFX and enable direct export.",
-                    ),
-                ),
-                (
-                    tr("menu.set_edit_targets", "Set Edit Targets"),
-                    self._show_post_load_prompt,
-                    "",
-                    tr(
-                        "menu.set_edit_targets.tip",
-                        "Choose which existing ideas/events files new content should be appended to.",
-                    ),
-                ),
-                (
-                    tr("menu.settings", "Settings"),
-                    self._open_settings,
-                    "",
-                    tr(
-                        "menu.settings.tip",
-                        "Configure mod path, GFX directories, MD detection, and UI options.",
-                    ),
-                ),
-            ],
-        )
-
-        # Right side of menu bar
-        self._errlog_btn = tk.Button(
-            menubar,
-            text=tr("menu.error_log", "Log"),
-            command=self._show_error_log,
-            bg="#080b10",
-            fg="#6e7681",
-            font=("Helvetica", 8, "bold"),
-            relief="flat",
-            padx=8,
-            pady=0,
-            cursor="hand2",
-            bd=0,
-            activebackground=BG_CARD,
-            activeforeground=TEXT,
-        )
-        self._errlog_btn.pack(side="right", padx=4)
-        Tooltip(
-            self._errlog_btn,
-            tr(
-                "menu.error_log.tip",
-                "View in-app error log.\nTurns red if any errors are caught during the session.",
-            ),
-        )
-        tk.Frame(menubar, bg=BORDER_G, width=1, height=16).pack(side="right", padx=2)
-        self._mod_lbl = tk.Label(
-            menubar,
-            text=tr("status.no_mod_loaded", "No mod loaded"),
-            bg="#080b10",
-            fg=TEXT_DIM,
-            font=("Helvetica", 8, "italic"),
-            padx=8,
-        )
-        self._mod_lbl.pack(side="right")
-        # close menu when clicking on canvas
-        self.bind(
-            "<Button-1>",
-            lambda e: _close_menu()
-            if self._open_menu_win[0]
-            and not str(e.widget).startswith(str(self._open_menu_win[0]))
-            else None,
-            add="+",
-        )
-
-        # hint label (used by _hint() method)
-        self._hint_lbl = tk.Label(
-            toolbar,
-            text=tr(
-                "hint.canvas_controls",
-                "Right-click canvas to place a focus  -  Ctrl+drag to pan  -  Scroll to zoom",
-            ),
-            bg=BG_DARK,
-            fg=TEXT_DIM,
-            font=("Helvetica", 8, "italic"),
-            anchor="w",
-            padx=10,
-        )
-        self._hint_lbl.pack(fill="x", pady=1)
-
-    def _build_toolbar_row2(self, toolbar):
-        """Build toolbar row 2: canvas action buttons and coord display."""
-        # ROW 2 — canvas tools (clean grouped toolbar)
-        row2 = tk.Frame(toolbar, bg="#090d14", height=36)
-        row2.pack(fill="x")
-        row2.pack_propagate(False)
-        self._conn_btn = None
-        self._mutex_btn = None
-
-        def _tb_sep():
-            tk.Frame(row2, bg=BORDER_G, width=1, height=20).pack(side="left", padx=5)
-
-        def _tb_lbl(t):
-            tk.Label(
-                row2,
-                text=t,
-                bg="#090d14",
-                fg=TEXT_DIM,
-                font=("Helvetica", 8, "bold"),
-                padx=4,
-            ).pack(side="left")
-
-        def _tb_btn(lbl, cmd, fg, bg, tip, padx=2):
-            b = tk.Button(
-                row2,
-                text=lbl,
-                command=cmd,
-                bg=bg,
-                fg=fg,
-                activebackground=BORDER_G,
-                activeforeground=TEXT,
-                font=("Helvetica", 9, "bold"),
-                relief="flat",
-                padx=9,
-                pady=2,
-                cursor="hand2",
-                highlightthickness=1,
-                highlightbackground=BORDER_G,
-            )
-            b.pack(side="left", padx=padx, pady=3)
-            Tooltip(b, tip)
-            return b
-
-        _tb_lbl(tr("toolbar.section.canvas", "Canvas"))
-        _tb_btn(
-            tr("toolbar.add_focus", "+ Focus"),
-            self._add_focus,
-            TEXT,
-            "#1e3a6e",
-            tr(
-                "toolbar.add_focus.tip",
-                "Add a new focus.\nAlso: right-click the canvas.",
-            ),
-        )
-        self._conn_btn = _tb_btn(
-            tr("toolbar.prereq", "Prereq"),
-            self._toggle_connect,
-            TEXT,
-            BG_CARD,
-            tr(
-                "toolbar.prereq.tip",
-                "Open the prerequisite picker for the selected focus.",
-            ),
-        )
-        self._mutex_btn = _tb_btn(
-            tr("toolbar.mutex", "Mutex"),
-            self._toggle_mutex,
-            ORANGE,
-            BG_CARD,
-            tr("toolbar.mutex.tip", "Draw a mutually exclusive link."),
-        )
-        _tb_sep()
-        _tb_lbl(tr("toolbar.section.tools", "Tools"))
-        _tb_btn(
-            tr("toolbar.ideas", "Ideas"),
-            self._national_spirit_wizard,
-            TEXT,
-            "#1a2040",
-            tr("toolbar.ideas.tip", "Build National Spirits / Ideas."),
-        )
-        _tb_btn(
-            tr("toolbar.dyn_mod", "Dyn Mod"),
-            self._dyn_mod_wizard,
-            TEXT,
-            BG_CARD,
-            tr("toolbar.dyn_mod.tip", "Dynamic Modifier wizard."),
-        )
-        _tb_btn(
-            tr("toolbar.decisions", "Decisions"),
-            self._decision_wizard,
-            TEXT,
-            BG_CARD,
-            tr("toolbar.decisions.tip", "Decision / Decision Category maker."),
-        )
-        _tb_btn(
-            tr("toolbar.events", "Events"),
-            self._event_wizard,
-            TEXT,
-            BG_CARD,
-            tr("toolbar.events.tip", "Event Maker wizard."),
-        )
-        _tb_btn(
-            tr("toolbar.add_income", "Add Income"),
-            self._additional_income_wizard,
-            "#4ade80",
-            "#0d2b1a",
-            tr(
-                "toolbar.add_income.tip",
-                "MD Additional Income Wizard - creates/links a spirit and wires up all money system files automatically.",
-            ),
-        )
-        _tb_sep()
-        _tb_lbl(tr("toolbar.section.select", "Select"))
-        self._msel_btn = _tb_btn(
-            tr("toolbar.multi", "Multi"),
-            self._toggle_multisel,
-            TEXT,
-            "#1a1a2e",
-            tr(
-                "toolbar.multi.tip",
-                "Toggle multi-select mode.\nCtrl+click focuses, then Delete.",
-            ),
-        )
-        _tb_btn(
-            tr("toolbar.delete_selected", "Del Selected"),
-            self._delete_selected,
-            TEXT,
-            "#4a1010",
-            tr("toolbar.delete_selected.tip", "Delete all selected focuses."),
-        )
-        _tb_sep()
-        _tb_btn(
-            tr("toolbar.clear_all", "Clear All"),
-            self._clear_all,
-            TEXT,
-            "#7f1d1d",
-            tr(
-                "toolbar.clear_all.tip",
-                "Delete ALL focuses from the canvas.\nSave first!",
-            ),
-        )
-        _tb_sep()
-        _tb_lbl(tr("toolbar.section.multi_tree", "Multi-Tree"))
-        _tb_btn(
-            tr("toolbar.shared", "+ Shared"),
-            lambda: self._load_extra_tree("shared"),
-            TEXT,
-            "#2d1c08",
-            tr(
-                "toolbar.shared.tip",
-                "Load a shared_focus tree alongside the main tree.\nIts focuses appear on canvas with amber [S] badges.",
-            ),
-        )
-        _tb_btn(
-            tr("toolbar.joint", "+ Joint"),
-            lambda: self._load_extra_tree("joint"),
-            TEXT,
-            "#1a0d40",
-            tr(
-                "toolbar.joint.tip",
-                "Load a joint_focus tree alongside the main tree.\nIts focuses appear on canvas with purple [J] badges.",
-            ),
-        )
-        _tb_btn(
-            tr("toolbar.load_all", "Load All"),
-            self._load_all_trees,
-            TEXT,
-            "#0d1a2e",
-            tr(
-                "toolbar.load_all.tip",
-                "Scan the mod's national_focus folder and load selected trees from a checklist.",
-            ),
-        )
-        _tb_btn(
-            tr("toolbar.save_all", "Save All"),
-            self._save_all_trees,
-            "#4ade80",
-            "#0d1a0a",
-            tr(
-                "toolbar.save_all.tip",
-                "Export all loaded trees (main + shared + joint) at once.",
-            ),
-        )
-
-        # Coord display right side
-        self._coord_lbl = tk.Label(
-            row2,
-            text="  x=0  y=0  ",
-            bg="#090d14",
-            fg=TEXT_DIM,
-            font=("Courier", 9),
-            padx=4,
-        )
-        self._coord_lbl.pack(side="right", padx=4)
-        tk.Label(
-            row2,
-            text=tr("toolbar.cursor", "Cursor:"),
-            bg="#090d14",
-            fg=TEXT_DIM,
-            font=("Helvetica", 8),
-        ).pack(side="right")
-
-        # Continuous Focus Position inputs (tree-level setting)
-        tk.Frame(row2, bg=BORDER_G, width=1, height=20).pack(side="right", padx=5)
-
-        def _cfp_commit(*_):
-            try:
-                self._cfp_x = int(self._cfp_x_var.get())
-            except Exception:
-                pass
-            try:
-                self._cfp_y = int(self._cfp_y_var.get())
-            except Exception:
-                pass
-
-        _cfp_y_ent = tk.Entry(
-            row2,
-            textvariable=self._cfp_y_var,
-            width=6,
-            bg=BG_CARD,
-            fg=TEXT,
-            insertbackground=BLUE,
-            font=("Courier", 9),
-            relief="flat",
-            highlightthickness=1,
-            highlightbackground=BORDER_G,
-        )
-        _cfp_y_ent.pack(side="right", ipady=1, padx=(0, 2))
-        _cfp_y_ent.bind("<FocusOut>", _cfp_commit)
-        _cfp_y_ent.bind("<Return>", _cfp_commit)
-        tk.Label(row2, text="y:", bg="#090d14", fg=TEXT_DIM, font=("Courier", 9)).pack(
-            side="right"
-        )
-        _cfp_x_ent = tk.Entry(
-            row2,
-            textvariable=self._cfp_x_var,
-            width=6,
-            bg=BG_CARD,
-            fg=TEXT,
-            insertbackground=BLUE,
-            font=("Courier", 9),
-            relief="flat",
-            highlightthickness=1,
-            highlightbackground=BORDER_G,
-        )
-        _cfp_x_ent.pack(side="right", ipady=1, padx=(0, 2))
-        _cfp_x_ent.bind("<FocusOut>", _cfp_commit)
-        _cfp_x_ent.bind("<Return>", _cfp_commit)
-        tk.Label(row2, text="x:", bg="#090d14", fg=TEXT_DIM, font=("Courier", 9)).pack(
-            side="right"
-        )
-        tk.Label(
-            row2,
-            text=tr("toolbar.continuous_focus_pos", "Continuous Focus Pos:"),
-            bg="#090d14",
-            fg=TEXT_DIM,
-            font=("Helvetica", 8),
-        ).pack(side="right")
 
     def _build_keybinds(self):
         """Bind all global keyboard shortcuts."""
@@ -2621,649 +1826,13 @@ class App(CanvasMixin, ModLoadingMixin, EffectsMixin, tk.Tk):
             self._redraw_now()
 
     def _open_gfx_browser(self):
+        """Delegate to the sidebar's focus-icon browser (ui/gfx_browser.py).
+
+        Scoped to gfx/interface/goals/, GFX_focus_-prefixed keys — see
+        open_focus_icon_browser's docstring for why this isn't the same
+        dialog the wizards use.
         """
-        Instant open: folder list shows immediately (no scanning).
-        Click folder -> file list built in background thread.
-        Scroll -> images loaded on demand per visible tile only.
-        """
-        if not MOD.loaded:
-            folder = filedialog.askdirectory(
-                title=tr(
-                    "filedialog.select_icon_folder", "Select folder with icon files"
-                )
-            )
-            if folder:
-                self._gfx_browse_files(folder)
-            return
-
-        goals_root = os.path.join(MOD.root, "gfx", "interface", "goals")
-        if not os.path.isdir(goals_root):
-            messagebox.showinfo(
-                tr("dialog.not_found.title", "Not Found"),
-                tr(
-                    "dialog.gfx_goals_not_found",
-                    "Could not find gfx/interface/goals/ in mod.",
-                ),
-            )
-            return
-
-        # ── Instant: just list directory names, no file counting ──
-        folders = []
-        # Loose files in goals/ root
-        loose = [
-            f
-            for f in os.listdir(goals_root)
-            if f.lower().endswith((".dds", ".png", ".tga"))
-        ]
-        if loose:
-            folders.append(("[goals root]", goals_root))
-        # Subfolders
-        for entry in sorted(os.listdir(goals_root)):
-            full = os.path.join(goals_root, entry)
-            if os.path.isdir(full):
-                folders.append((entry, full))
-        if not folders:
-            messagebox.showinfo(
-                tr("dialog.no_folders.title", "No Folders"),
-                tr(
-                    "dialog.no_goal_subfolders",
-                    "No subfolders found in gfx/interface/goals/",
-                ),
-            )
-            return
-
-        # ── Window ────────────────────────────────────────────────
-        win = tk.Toplevel(self)
-        win.title(tr("gfx.browser.title", "GFX Browser"))
-        win.configure(bg=BG_DARK)
-        win.geometry("900x580")
-        win.resizable(True, True)
-
-        panes = tk.Frame(win, bg=BG_DARK)
-        panes.pack(fill="both", expand=True, padx=8, pady=8)
-
-        # Left: folder list (text only, instant)
-        lf = tk.Frame(panes, bg=BG_PANEL, width=200)
-        lf.pack(side="left", fill="y", padx=(0, 6))
-        lf.pack_propagate(False)
-        tk.Label(
-            lf,
-            text=tr("gfx.folders", "  FOLDERS"),
-            bg=BG_PANEL,
-            fg=TEXT_DIM,
-            font=("Helvetica", 9, "bold"),
-            anchor="w",
-            pady=6,
-        ).pack(fill="x")
-        tk.Frame(lf, bg=BORDER_G, height=1).pack(fill="x")
-        folder_lb = tk.Listbox(
-            lf,
-            bg=BG_CARD,
-            fg=TEXT,
-            selectbackground=BLUE,
-            selectforeground=TEXT,
-            font=("Courier", 9),
-            relief="flat",
-            bd=0,
-            activestyle="none",
-            highlightthickness=0,
-        )
-        fsb = tk.Scrollbar(lf, orient="vertical", command=folder_lb.yview)
-        folder_lb.configure(yscrollcommand=fsb.set)
-        fsb.pack(side="right", fill="y")
-        folder_lb.pack(fill="both", expand=True, padx=2, pady=4)
-        for display, _ in folders:
-            folder_lb.insert("end", "  " + display)
-
-        # Right panel
-        rf = tk.Frame(panes, bg=BG_DARK)
-        rf.pack(side="left", fill="both", expand=True)
-
-        top_r = tk.Frame(rf, bg=BG_DARK)
-        top_r.pack(fill="x", pady=(0, 6))
-        tk.Label(
-            top_r,
-            text=tr("gfx.filter", "Filter:"),
-            bg=BG_DARK,
-            fg=TEXT_DIM,
-            font=("Helvetica", 9),
-        ).pack(side="left")
-        search_var = tk.StringVar()
-        tk.Entry(
-            top_r,
-            textvariable=search_var,
-            bg=BG_CARD,
-            fg=TEXT,
-            insertbackground=BLUE,
-            font=("Helvetica", 10),
-            relief="flat",
-            highlightthickness=1,
-            highlightbackground=BORDER_G,
-        ).pack(side="left", padx=6, fill="x", expand=True, ipady=3)
-        status_lbl = tk.Label(
-            top_r,
-            text=tr("gfx.select_folder", "select a folder"),
-            bg=BG_DARK,
-            fg=TEXT_DIM,
-            font=("Helvetica", 9),
-        )
-        status_lbl.pack(side="right", padx=6)
-
-        cv_frame = tk.Frame(rf, bg=BG_PANEL)
-        cv_frame.pack(fill="both", expand=True)
-        cv = tk.Canvas(cv_frame, bg=BG_PANEL, highlightthickness=0)
-        vsb = tk.Scrollbar(cv_frame, orient="vertical", command=cv.yview)
-        cv.configure(yscrollcommand=vsb.set)
-        vsb.pack(side="right", fill="y")
-        cv.pack(side="left", fill="both", expand=True)
-
-        # Bottom bar
-        bot = tk.Frame(win, bg=BG_DARK)
-        bot.pack(fill="x", padx=10, pady=6)
-        selected_var = tk.StringVar(value="")
-        _initial_gfx = self._fv_gfx.get()
-        tk.Label(
-            bot, textvariable=selected_var, bg=BG_DARK, fg=BLUE, font=("Helvetica", 9)
-        ).pack(side="left", padx=4)
-        tk.Button(
-            bot,
-            text=tr("common.cancel", "Cancel"),
-            command=win.destroy,
-            bg=BG_CARD,
-            fg=TEXT,
-            relief="flat",
-            font=("Helvetica", 9),
-            padx=10,
-            pady=4,
-            cursor="hand2",
-        ).pack(side="right", padx=4)
-
-        def _apply():
-            self._set_gfx(selected_var.get())
-            win.destroy()
-
-        _sel_btn = tk.Button(
-            bot,
-            text=tr("common.select", "Select"),
-            command=_apply,
-            bg="#1a3322",
-            fg="#4b7a5e",
-            relief="flat",
-            font=("Helvetica", 10, "bold"),
-            padx=14,
-            pady=5,
-            cursor="arrow",
-            state="disabled",
-        )
-        _sel_btn.pack(side="right")
-
-        def _on_sel_change(*_):
-            v = selected_var.get()
-            if v and v != _initial_gfx:
-                _sel_btn.config(
-                    bg="#14532d", fg="#0a0a0a", cursor="hand2", state="normal"
-                )
-            elif v:
-                _sel_btn.config(
-                    bg="#1e6b3a", fg="#c8f0d8", cursor="hand2", state="normal"
-                )
-            else:
-                _sel_btn.config(
-                    bg="#1a3322", fg="#4b7a5e", cursor="arrow", state="disabled"
-                )
-
-        selected_var.trace_add("write", _on_sel_change)
-
-        # ── Grid constants ────────────────────────────────────────
-        COLS = 5
-        TILE_W = 110
-        TILE_H = 100
-        PAD = 6
-
-        # ── State ─────────────────────────────────────────────────
-        _st = {
-            "pairs": [],  # [(gfx_key, path), ...]
-            "img_cache": {},  # path -> PhotoImage | None
-            "drawn": set(),  # indices already rendered
-            "canvas_ids": {},  # idx -> (rect_id, img_id or txt_id, lbl_id)
-            "sel_idx": None,
-            "scan_job": None,  # pending after() job
-            "load_gen": None,  # background thread
-        }
-
-        def _tile_xy(idx):
-            col = idx % COLS
-            row = idx // COLS
-            return PAD + col * (TILE_W + PAD), PAD + row * (TILE_H + PAD)
-
-        def _select_tile(idx):
-            old = _st["sel_idx"]
-            if old is not None and old in _st["canvas_ids"]:
-                rid, _, _ = _st["canvas_ids"][old]
-                cv.itemconfig(rid, fill=BG_CARD, outline=BORDER_G)
-            _st["sel_idx"] = idx
-            gfx_key = _st["pairs"][idx][0]
-            selected_var.set(gfx_key)
-            if idx in _st["canvas_ids"]:
-                rid, _, _ = _st["canvas_ids"][idx]
-                cv.itemconfig(rid, fill=SEL_BG, outline=BLUE)
-
-        def _draw_tile(idx):
-            """Draw placeholder immediately; image filled in by background thread."""
-            if idx in _st["drawn"]:
-                return
-            _st["drawn"].add(idx)
-            gfx_key, path = _st["pairs"][idx]
-            x, y = _tile_xy(idx)
-            is_sel = gfx_key == selected_var.get()
-            rid = cv.create_rectangle(
-                x,
-                y,
-                x + TILE_W,
-                y + TILE_H,
-                fill=SEL_BG if is_sel else BG_CARD,
-                outline=BLUE if is_sel else BORDER_G,
-                width=2,
-                tags=("tile", "t%d" % idx),
-            )
-            # Placeholder spinner text while image loads
-            iid = cv.create_text(
-                x + TILE_W // 2,
-                y + 44,
-                text="...",
-                fill=TEXT_DIM,
-                font=("Helvetica", 14),
-                tags=("tile", "t%d" % idx),
-            )
-            short = gfx_key.replace("GFX_focus_", "").replace("GFX_goal_", "")
-            short = (short[:16] + "...") if len(short) > 16 else short
-            lid = cv.create_text(
-                x + TILE_W // 2,
-                y + TILE_H - 14,
-                text=short,
-                fill=TEXT_DIM,
-                font=("Helvetica", 7),
-                width=TILE_W - 8,
-                tags=("tile", "t%d" % idx),
-            )
-            _st["canvas_ids"][idx] = (rid, iid, lid)
-            for item in (rid, iid, lid):
-                cv.tag_bind(item, "<Button-1>", lambda e, i=idx: _select_tile(i))
-                cv.tag_bind(
-                    item,
-                    "<Double-Button-1>",
-                    lambda e, i=idx: [_select_tile(i), _apply()],
-                )
-            # If image already cached, fill it in right now
-            if path in _st["img_cache"]:
-                _fill_image(idx)
-
-        def _fill_image(idx):
-            """Replace placeholder with actual image (called from main thread)."""
-            if idx not in _st["canvas_ids"]:
-                return
-            rid, iid, lid = _st["canvas_ids"][idx]
-            gfx_key, path = _st["pairs"][idx]
-            img = _st["img_cache"].get(path)
-            cv.delete(iid)
-            if img:
-                new_iid = cv.create_image(
-                    _tile_xy(idx)[0] + TILE_W // 2,
-                    _tile_xy(idx)[1] + 44,
-                    anchor="center",
-                    image=img,
-                    tags=("tile", "t%d" % idx),
-                )
-            else:
-                new_iid = cv.create_text(
-                    _tile_xy(idx)[0] + TILE_W // 2,
-                    _tile_xy(idx)[1] + 30,
-                    text="?",
-                    fill=TEXT_DIM,
-                    font=("Helvetica", 20),
-                    tags=("tile", "t%d" % idx),
-                )
-            _st["canvas_ids"][idx] = (rid, new_iid, lid)
-            for item in (rid, new_iid, lid):
-                cv.tag_bind(item, "<Button-1>", lambda e, i=idx: _select_tile(i))
-                cv.tag_bind(
-                    item,
-                    "<Double-Button-1>",
-                    lambda e, i=idx: [_select_tile(i), _apply()],
-                )
-
-        def _bg_load_images(pairs_snapshot, indices):
-            """Background: load images for given indices, post update to main thread."""
-            for idx in indices:
-                if idx >= len(pairs_snapshot):
-                    break
-                gfx_key, path = pairs_snapshot[idx]
-                if path in _st["img_cache"]:
-                    continue
-                img = None
-                try:
-                    if _PIL_OK and os.path.exists(path):
-                        pil = _PILImage.open(path).convert("RGBA")
-                        rs = getattr(
-                            _PILImage, "LANCZOS", getattr(_PILImage, "ANTIALIAS", 1)
-                        )
-                        pil = pil.resize((72, 72), rs)
-                        img = _PILImageTk.PhotoImage(pil)
-                except Exception:
-                    pass
-                _st["img_cache"][path] = img
-                # Tell main thread to update this tile
-                _safe_after(win, 0, lambda i=idx: _fill_image(i))
-
-        def _lazy_fill(*_):
-            """Draw placeholders for visible tiles; kick off background image load."""
-            if not _st["pairs"]:
-                return
-            cv.update_idletasks()
-            top = cv.canvasy(0)
-            bottom = cv.canvasy(cv.winfo_height())
-            visible = []
-            for idx in range(len(_st["pairs"])):
-                _, ty = _tile_xy(idx)
-                if ty + TILE_H >= top and ty <= bottom:
-                    _draw_tile(idx)
-                    visible.append(idx)
-            # Background load for visible + next ~40 tiles
-            last = max(visible) if visible else 0
-            ahead = list(range(last + 1, min(last + 41, len(_st["pairs"]))))
-            to_load = [
-                i
-                for i in (visible + ahead)
-                if _st["pairs"][i][1] not in _st["img_cache"]
-            ]
-            if to_load:
-                snapshot = list(_st["pairs"])
-                t = threading.Thread(
-                    target=_bg_load_images, args=(snapshot, to_load), daemon=True
-                )
-                t.start()
-
-        def _rebuild(pairs):
-            cv.delete("all")
-            _st["pairs"] = pairs
-            _st["drawn"].clear()
-            _st["canvas_ids"].clear()
-            _st["sel_idx"] = None
-            # Keep img_cache across folders — avoids reloading same files
-            if not pairs:
-                status_lbl.config(text=tr("gfx.icons_count", "{count} icons", count=0))
-                return
-            status_lbl.config(text="%d icons" % len(pairs))
-            rows = (len(pairs) + COLS - 1) // COLS
-            total_h = PAD + rows * (TILE_H + PAD)
-            total_w = PAD + COLS * (TILE_W + PAD)
-            cv.configure(scrollregion=(0, 0, total_w, total_h))
-            cv.yview_moveto(0)
-            _safe_after_idle(win, _lazy_fill)
-
-        def _collect_files(folder_path):
-            """Fast file scan — no MOD.sprites lookup, just walk the folder."""
-            pairs = []
-            ft = search_var.get().lower()
-            for root_d, dirs, fnames in os.walk(folder_path):
-                dirs.sort()
-                for fname in sorted(fnames):
-                    if not fname.lower().endswith((".dds", ".png", ".tga")):
-                        continue
-                    if ft and ft not in fname.lower():
-                        continue
-                    full = os.path.join(root_d, fname)
-                    stem = os.path.splitext(fname)[0]
-                    gfx_key = "GFX_focus_" + stem
-                    pairs.append((gfx_key, full))
-            return pairs
-
-        def _load_folder(folder_path):
-            status_lbl.config(text=tr("gfx.scanning", "scanning..."))
-            win.update_idletasks()
-            # Collect file list (fast — just filenames, no image loading)
-            pairs = _collect_files(folder_path)
-            _rebuild(pairs)
-
-        def _on_folder_select(evt):
-            sel = folder_lb.curselection()
-            if not sel:
-                return
-            _load_folder(folders[sel[0]][1])
-
-        cv.bind("<Configure>", lambda e: _safe_after_idle(win, _lazy_fill))
-        for ev in ("<MouseWheel>", "<Button-4>", "<Button-5>"):
-            cv.bind(
-                ev,
-                lambda e: [
-                    cv.yview_scroll(-1 if (e.delta > 0 or e.num == 4) else 1, "units"),
-                    _safe_after_idle(win, _lazy_fill),
-                ],
-            )
-        folder_lb.bind("<<ListboxSelect>>", _on_folder_select)
-        search_var.trace_add(
-            "write",
-            lambda *_: _safe_after(
-                win,
-                300,
-                lambda: _on_folder_select(None) if folder_lb.curselection() else None,
-            ),
-        )
-
-        # Auto-select first folder
-        if folders:
-            folder_lb.selection_set(0)
-            _load_folder(folders[0][1])
-
-    def _gfx_browse_files(self, folder):
-        """Fallback lazy browser — no mod loaded."""
-        all_files = sorted(
-            (f, os.path.join(folder, f))
-            for f in os.listdir(folder)
-            if f.lower().endswith((".dds", ".png", ".tga"))
-        )
-        if not all_files:
-            messagebox.showinfo(
-                tr("dialog.no_files.title", "No Files"),
-                tr("dialog.no_gfx_files", "No .dds/.png/.tga files found."),
-            )
-            return
-        pairs = [("GFX_focus_" + os.path.splitext(f)[0], p) for f, p in all_files]
-        COLS = 5
-        TILE_W = 110
-        TILE_H = 100
-        PAD = 6
-        win = tk.Toplevel(self)
-        win.title(tr("gfx.browser.title", "GFX Browser"))
-        win.configure(bg=BG_DARK)
-        win.geometry("700x480")
-        cvf = tk.Frame(win, bg=BG_PANEL)
-        cvf.pack(fill="both", expand=True, padx=8, pady=8)
-        cv = tk.Canvas(cvf, bg=BG_PANEL, highlightthickness=0)
-        vsb = tk.Scrollbar(cvf, orient="vertical", command=cv.yview)
-        cv.configure(yscrollcommand=vsb.set)
-        vsb.pack(side="right", fill="y")
-        cv.pack(fill="both", expand=True)
-        rows = (len(pairs) + COLS - 1) // COLS
-        cv.configure(
-            scrollregion=(
-                0,
-                0,
-                PAD + COLS * (TILE_W + PAD),
-                PAD + rows * (TILE_H + PAD),
-            )
-        )
-        selected_var = tk.StringVar(value=self._fv_gfx.get())
-        bot = tk.Frame(win, bg=BG_DARK)
-        bot.pack(fill="x", padx=8, pady=6)
-        tk.Label(
-            bot, textvariable=selected_var, bg=BG_DARK, fg=BLUE, font=("Helvetica", 9)
-        ).pack(side="left")
-        tk.Button(
-            bot,
-            text=tr("common.cancel", "Cancel"),
-            command=win.destroy,
-            bg=BG_CARD,
-            fg=TEXT,
-            relief="flat",
-            font=("Helvetica", 9),
-            padx=10,
-            pady=4,
-            cursor="hand2",
-        ).pack(side="right", padx=4)
-
-        def _apply():
-            self._set_gfx(selected_var.get())
-            win.destroy()
-
-        tk.Button(
-            bot,
-            text=tr("common.select", "Select"),
-            command=_apply,
-            bg="#14532d",
-            fg="#0a0a0a",
-            relief="flat",
-            font=("Helvetica", 10, "bold"),
-            padx=14,
-            pady=5,
-            cursor="hand2",
-        ).pack(side="right")
-        _cache = {}
-        _drawn = set()
-        _ids = {}
-
-        def _txy(i):
-            return PAD + (i % COLS) * (TILE_W + PAD), PAD + (i // COLS) * (TILE_H + PAD)
-
-        def _fill(idx):
-            if idx not in _ids:
-                return
-            rid, iid, lid = _ids[idx]
-            gfx_key, path = pairs[idx]
-            img = _cache.get(path)
-            cv.delete(iid)
-            x, y = _txy(idx)
-            if img:
-                n = cv.create_image(
-                    x + TILE_W // 2, y + 44, anchor="center", image=img, tags="tile"
-                )
-            else:
-                n = cv.create_text(
-                    x + TILE_W // 2,
-                    y + 30,
-                    text="?",
-                    fill=TEXT_DIM,
-                    font=("Helvetica", 20),
-                    tags="tile",
-                )
-            _ids[idx] = (rid, n, lid)
-            for item in (rid, n, lid):
-                cv.tag_bind(
-                    item, "<Button-1>", lambda e, k=gfx_key: selected_var.set(k)
-                )
-                cv.tag_bind(
-                    item,
-                    "<Double-Button-1>",
-                    lambda e, k=gfx_key: [selected_var.set(k), _apply()],
-                )
-
-        def _draw(idx):
-            if idx in _drawn:
-                return
-            _drawn.add(idx)
-            gfx_key, path = pairs[idx]
-            x, y = _txy(idx)
-            rid = cv.create_rectangle(
-                x,
-                y,
-                x + TILE_W,
-                y + TILE_H,
-                fill=BG_CARD,
-                outline=BORDER_G,
-                width=2,
-                tags="tile",
-            )
-            iid = cv.create_text(
-                x + TILE_W // 2,
-                y + 44,
-                text="...",
-                fill=TEXT_DIM,
-                font=("Helvetica", 14),
-                tags="tile",
-            )
-            short = gfx_key.replace("GFX_focus_", "")
-            short = (short[:16] + "...") if len(short) > 16 else short
-            lid = cv.create_text(
-                x + TILE_W // 2,
-                y + TILE_H - 14,
-                text=short,
-                fill=TEXT_DIM,
-                font=("Helvetica", 7),
-                width=TILE_W - 8,
-                tags="tile",
-            )
-            _ids[idx] = (rid, iid, lid)
-            if path in _cache:
-                _fill(idx)
-            for item in (rid, iid, lid):
-                cv.tag_bind(
-                    item, "<Button-1>", lambda e, k=gfx_key: selected_var.set(k)
-                )
-                cv.tag_bind(
-                    item,
-                    "<Double-Button-1>",
-                    lambda e, k=gfx_key: [selected_var.set(k), _apply()],
-                )
-
-        def _bg(snap, idxs):
-            for i in idxs:
-                if i >= len(snap):
-                    break
-                _, path = snap[i]
-                if path in _cache:
-                    continue
-                img = None
-                try:
-                    if _PIL_OK and os.path.exists(path):
-                        pil = _PILImage.open(path).convert("RGBA")
-                        rs = getattr(
-                            _PILImage, "LANCZOS", getattr(_PILImage, "ANTIALIAS", 1)
-                        )
-                        pil = pil.resize((72, 72), rs)
-                        img = _PILImageTk.PhotoImage(pil)
-                except Exception:
-                    pass
-                _cache[path] = img
-                _safe_after(win, 0, lambda x=i: _fill(x))
-
-        def _lazy(*_):
-            top = cv.canvasy(0)
-            bot_y = cv.canvasy(cv.winfo_height())
-            vis = []
-            for i in range(len(pairs)):
-                _, ty = _txy(i)
-                if ty + TILE_H >= top and ty <= bot_y:
-                    _draw(i)
-                    vis.append(i)
-            last = max(vis) if vis else 0
-            ahead = list(range(last + 1, min(last + 41, len(pairs))))
-            to_load = [i for i in vis + ahead if pairs[i][1] not in _cache]
-            if to_load:
-                t = threading.Thread(
-                    target=_bg, args=(list(pairs), to_load), daemon=True
-                )
-                t.start()
-
-        cv.bind("<Configure>", lambda e: _safe_after_idle(win, _lazy))
-        for ev in ("<MouseWheel>", "<Button-4>", "<Button-5>"):
-            cv.bind(
-                ev,
-                lambda e: [
-                    cv.yview_scroll(-1 if (e.delta > 0 or e.num == 4) else 1, "units"),
-                    _safe_after_idle(win, _lazy),
-                ],
-            )
-        _safe_after_idle(win, _lazy)
+        open_focus_icon_browser(self, self._set_gfx, self._fv_gfx.get())
 
     def _attach_autocomplete(self, entry_widget, var, get_choices_fn):
         """Attach a live autocomplete popup to any Entry widget."""
