+++ tests/test_ui_smoke.py
@@ -14,6 +14,36 @@ def test_border_is_exported_from_hoi4cm_ui():
     assert BORDER.startswith("#")
 
 
+def test_gfx_browser_and_settings_importable_from_facade():
+    """hoi4cm.ui re-exports the dialogs the monolith wires in directly.
+
+    No Tk objects get created here — just checking the names resolve, same
+    headless guarantee as the BORDER check above.
+    """
+    from hoi4cm.ui import open_focus_icon_browser, open_settings
+
+    assert callable(open_focus_icon_browser)
+    assert callable(open_settings)
+
+
+def test_menubar_and_toolbar_import_headless():
+    """menubar.py/toolbar.py are one-shot Tk builders, not testable end to end.
+
+    This only proves the module bodies parse and their top-level definitions
+    (constant imports, function signature) don't blow up on import with no
+    display — no Toplevel/Tk() gets created here. Building a real menu bar
+    needs a live App and is manual-only (see testing.md).
+    """
+    from hoi4cm.ui import build_menubar, build_toolbar_row2
+    from hoi4cm.ui.menubar import build_menubar as build_menubar_direct
+    from hoi4cm.ui.toolbar import build_toolbar_row2 as build_toolbar_row2_direct
+
+    assert build_menubar is build_menubar_direct
+    assert build_toolbar_row2 is build_toolbar_row2_direct
+    assert callable(build_menubar)
+    assert callable(build_toolbar_row2)
+
+
 def test_all_theme_constants_are_strings():
     """Every colour constant exported by hoi4cm.ui.theme should be a hex string."""
     import hoi4cm.ui.theme as theme
